+++ example/android/app/src/main/AndroidManifest.xml
@@ -34,9 +34,7 @@
         <meta-data
             android:name="flutterEmbedding"
             android:value="2" />
-        <meta-data
-    android:name="io.flutter.embedding.android.EnableImpeller"
-    android:value="false" />
+
     </application>
     <!-- Required to query activities that can process text, see:
          https://developer.android.com/training/package-visibility and
