+++ lib/src/decoder/qr_code_dart_scan_decoder.dart
@@ -1,4 +1,5 @@
 import 'package:camera/camera.dart';
+import 'package:flutter/foundation.dart';
 import 'package:qr_code_dart_decoder/qr_code_dart_decoder.dart';
 import 'package:qr_code_dart_scan/src/util/image_decode_orientation.dart';
 
@@ -46,6 +47,7 @@ class QRCodeDartScanDecoder {
     _isolateDecoder = IsolateDecoder(
       formats: formats,
       countIsolates: countIsolates,
+      preYuvProcessor: CropBackgroundYuvProcessor(debug: kDebugMode),
     )..start();
   }
